Update of version in .vsixmanifest
Changed the version number for the extension 

The InstallationTarget element in the .vsixmanifest file is specifically targeting Visual Studio version 16.0 and not any versions higher than 16.0 on x86 architecture. Therefore, if you have a higher version of Visual Studio installed, such as Visual Studio Professional 2019 version 16.11.23, the extension may not be compatible.